Support for the class directive.
If any object has a "classes" attribute, the first class will be looked up in the stylesheet from styles.py and
used instead of what would regularly be used.



So, for example, if you have this in styles.py:

    stylesheet.add(ParagraphStyle(name='special',
                                  parent=stylesheet['Normal'],
                                  fontName=stdMono,
                                  bulletFontName=stdMono,
                                  fontSize=12,
                                  bulletFontSize=8,
                                  leading=8.8,
                                  firstLineIndent=0,
                                  leftIndent=12,
                                  spaceBefore=4))

This is what would happen:

------------

This paragraph is normal.

.. class:: special

This paragraph is 12pt monospaced.

This one is normal, too.



git-svn-id: http://rst2pdf.googlecode.com/svn/trunk@7 3777fadb-0f44-0410-9e7f-9d8fa6171d72