Support for inline text styling using custom roles.
For example, suppose you want the word "red" to be shown red, you would define a
redtext style and use this:
  
  
.. role:: redtext

This is :redtext:`red`.

The bad news is that to define custom styles right now you need to edit styles.py, 
but don't worry, that's getting fixed very very soon.



git-svn-id: http://rst2pdf.googlecode.com/svn/trunk@18 3777fadb-0f44-0410-9e7f-9d8fa6171d72