Added temporary code to autotest to monkeypatch ReportLab 2.4.
The purpose of the code is to be able to distinguish between
real bugs and errors in the RL2.3 -> 2.4 transition, and differences
that just don't matter.

In the monkeypatch itself, there is one difference that is compensated
for that requires further investigation (paraFontSizeHeightOffset), and
several that are irrelevant to actual content production.

After applying the monkeypatch, 95 of the tests pass with 'good'
checksums that are identical to the RL 2.3 checksums.



git-svn-id: http://rst2pdf.googlecode.com/svn/trunk@1628 3777fadb-0f44-0410-9e7f-9d8fa6171d72