Modified header/footer code to allow page templates to turn off headers and footers, even when header/footer information is available (via showHeader/showFooter).
This functionality was in there, but not working properly,
I think.

However, this fix changed the result of test_issue_199.

When I investigated, though, it seemed (looking at the logic
in flowables) that the new behavior might be expected/desired,
because there is some code to deliberately disable header/footer
for blank pages (which now happens on this test).

Will require guidance from Roberto on whether this fix is
correct, or if something else should be done.



git-svn-id: http://rst2pdf.googlecode.com/svn/trunk@1706 3777fadb-0f44-0410-9e7f-9d8fa6171d72